#24 Finalized checkpoints/respawn
Finalised dying and checkpoints, both require the presence of a 'respawn manager' game object. Both respawn and checkpoints are handled within 'DeathCheckpoint' script.
When applying this script, only thing neccesary is ensuring a box collider is applied + set to trigger, and that isDeathboth bool is set to true or false.